fix: consolidate process chain formatting and review fixes
- Extract shared FormatChain/ShortenProcessName into ProcessChainEntry
- HealthReportFormatter and Program.Command delegate to shared helper
- DiscoveryOutputFormatter uses ShortenProcessName (keeps Spectre markup)
- Fix ToInt32() overflow on 64-bit: use (int)IntPtr cast instead
- Fix test assertion for new chain format (reversed, shortened)
- Replace IDE-specific name in test with generic "ide-process"
